requirements.txt
pillow>=8.2.0,<8.3.0
